refeat: Simplify GerarOrcamentoScreen code
